requirements.txt
# Core ML/DL Libraries
torch
torchvision
torchaudio

# Data Handling
pandas
numpy
scikit-learn

# Experiment Tracking
mlflow

# API/Serving
fastapi
uvicorn[standard]

# Plotting & Image Handling
matplotlib
seaborn
Pillow

# Code Quality
ruff
black